refactor(cli): move presentation helpers (style, output, problem_json) out of commands into a cli module
