+++ src/cli/output.rs
