+++ src/commands/apply.rs
@@ -1,4 +1,4 @@
-use crate::commands::problem_json::render_response_error;
+use crate::cli::problem_json::render_response_error;
 use crate::config::config::{Config, get_config_dir, load_auth_config};
 use crate::exit_code;
 use crate::models::deployments::EnvValue;
