+++ src/commands/deployment/delete.rs
@@ -3,8 +3,8 @@ use clap::ArgAction;
 use clap::ArgMatches;
 use clap::Command;
 
-use crate::commands::problem_json::http_error;
-use crate::commands::style;
+use crate::cli::problem_json::http_error;
+use crate::cli::style;
 use crate::config::config::Config;
 use crate::config::config::load_auth_config;
 use crate::exit_code;
