+++ src/commands/deployment/list.rs
@@ -1,7 +1,7 @@
 use crate::api::dto::deployment::DeploymentOutput;
-use crate::commands::output::{output_arg, output_format};
-use crate::commands::problem_json::http_error_list;
-use crate::commands::style;
+use crate::cli::output::{output_arg, output_format};
+use crate::cli::problem_json::http_error_list;
+use crate::cli::style;
 use crate::config::config::Config;
 use crate::config::config::load_auth_config;
 use crate::exit_code;
