+++ src/commands/secret/delete.rs
@@ -4,8 +4,8 @@ use clap::Command;
 use serde::Deserialize;
 use std::io::{self, Write};
 
-use crate::commands::problem_json::http_error;
-use crate::commands::style;
+use crate::cli::problem_json::http_error;
+use crate::cli::style;
 use crate::config::config::Config;
 use crate::config::config::load_auth_config;
 use crate::exit_code;
