+++ src/commands/user/update.rs
@@ -1,6 +1,6 @@
 use crate::api::dto::user::UserOutput;
-use crate::commands::problem_json::{http_error, render_response_error};
-use crate::commands::style;
+use crate::cli::problem_json::{http_error, render_response_error};
+use crate::cli::style;
 use crate::config::config::Config;
 use crate::config::config::load_auth_config;
 use crate::exit_code;
